requirements.txt
pymongo==3.8.0
python-dotenv==0.10.3
redis==3.3.11
Flask==1.1.1
